refactor(frontend): final structure cleanup
- Move ResultsPage.tsx to pages/ (it's a routed page, not a component)
- Rename features/home/Carousel.tsx → ModeCarousel.tsx (was too generic)
- Delete WebSocketDebug.tsx (unused dead code)
