+++ srcs/frontend/src/features/home/ModeCarousel.tsx
