+++ srcs/frontend/src/pages/ResultsPage.tsx
